Add basic UCI commands
The commands 'isready', 'uci', and 'quit' were implemented into the
engine. The 'isready' command is used as a ping command. If the engine
is doing something, the gui can send this command to check if the engine
is still running. The 'uci' command is used to initialize the engine and
give information about the engine to the gui. The 'quit' command simply
ends the program.