Define some board features
A board struct was created. This struct contained two sets of bitboards.
The first set contained all the bitboards representing the pieces by
type and the second represented the pieces by side. One bitboard of each
set can be bitwise ANDed together to get a bitboard representing a piece
for a certain player.